Abstracted trains as individual component
